+++ plone/app/users/setuphandlers.py
@@ -9,12 +9,12 @@ FILE = 'userschema.xml'
 
 
 def import_schema(context):
-    """Import TTW Schema """
+    """Import TTW Schema
+    """
     data = context.readDataFile(FILE)
     if data is None:
         return
     ttw.applySchema(data)
-
     logger.info('Imported schema')
 
 
@@ -22,5 +22,5 @@ def export_schema(context):
     """Export TTW schema
     """
     schema = ttw.serialize_ttw_schema()
-    logger.info('Exported schema')
     context.writeDataFile(FILE, schema, 'text/xml')
+    logger.info('Exported schema')